Move relations into mapper namespace
