+++ src/Orm/Relations/BelongsTo.php
@@ -7,26 +7,25 @@
  * file that was distributed with this source code.
  *
  * Date: 01.02.2016
- * Time: 15:54
+ * Time: 15:18
  *
  */
 
-namespace Blast\Db\Relations;
+namespace Blast\Db\Orm\Relations;
 
 
 use Blast\Db\Entity\EntityInterface;
 use Blast\Db\Query;
 
-class HasOne extends AbstractRelation
+class BelongsTo extends AbstractRelation
 {
     /**
      * BelongsTo constructor.
      * @param EntityInterface $model current entity instance
      * @param string|EntityInterface $foreignEntity Entity which have a key field in current entity
      * @param string|integer|null $localKey Field name on current entity which matches up with foreign key of foreign entity
-     * @param string|integer|null $foreignKey name on foreign entity which matches up with local key of current entity
      */
-    public function __construct(EntityInterface $model, $foreignEntity, $localKey = null, $foreignKey = null)
+    public function __construct(EntityInterface $model, $foreignEntity, $localKey = null)
     {
         //local config
         $this->entity = $model;
@@ -34,12 +33,10 @@ class HasOne extends AbstractRelation
 
         //foreign config
         $this->foreignEntity = $foreignEntity;
-        $this->foreignKey = $foreignKey;
     }
 
     /**
      * Save foreign entity and store value of foreign key into local key field
-     *
      * @return EntityInterface
      */
     public function save()
@@ -61,8 +58,9 @@ class HasOne extends AbstractRelation
     {
         $query = $this->getForeignEntity()->getMapper()->select();
         $result = $query->where(
-            $query->expr()->eq($this->getLocalKey(), $this->getForeignEntity()->get($this->getLocalKey()))
+            $query->expr()->eq($this->getForeignKey(), $this->getEntity()->get($this->getLocalKey()))
         )->setMaxResults(1)->execute(Query::RESULT_ENTITY);
+
         return $this->setResults($result)->getResults();
     }
