+++ src/Orm/Relations/HasMany.php
@@ -6,26 +6,34 @@
  * For the full copyright and license information, please view the LICENSE.txt
  * file that was distributed with this source code.
  *
- * Date: 01.02.2016
- * Time: 15:18
+ * Date: 02.02.2016
+ * Time: 15:56
  *
  */
 
-namespace Blast\Db\Relations;
+namespace Blast\Db\Orm\Relations;
 
 
+use Blast\Db\Entity\CollectionInterface;
 use Blast\Db\Entity\EntityInterface;
 use Blast\Db\Query;
 
-class BelongsTo extends AbstractRelation
+class HasMany extends AbstractRelation
 {
+
+    /**
+     * @var EntityInterface[]
+     */
+    protected $foreignEntities = [];
+
     /**
      * BelongsTo constructor.
      * @param EntityInterface $model current entity instance
      * @param string|EntityInterface $foreignEntity Entity which have a key field in current entity
      * @param string|integer|null $localKey Field name on current entity which matches up with foreign key of foreign entity
+     * @param string|integer|null $foreignKey name on foreign entity which matches up with local key of current entity
      */
-    public function __construct(EntityInterface $model, $foreignEntity, $localKey = null)
+    public function __construct(EntityInterface $model, $foreignEntity, $localKey = null, $foreignKey = null)
     {
         //local config
         $this->entity = $model;
@@ -33,9 +41,27 @@ class BelongsTo extends AbstractRelation
 
         //foreign config
         $this->foreignEntity = $foreignEntity;
+        $this->foreignKey = $foreignKey;
     }
 
     /**
+     * @return \Blast\Db\Entity\EntityInterface[]
+     */
+    public function getForeignEntities()
+    {
+        return $this->foreignEntities;
+    }
+
+    /**
+     * @param \Blast\Db\Entity\EntityInterface[] $foreignEntities
+     */
+    public function setForeignEntities($foreignEntities)
+    {
+        $this->foreignEntities = $foreignEntities;
+    }
+
+    /**
+     * @todo Entity should be able to save many entities
      * Save foreign entity and store value of foreign key into local key field
      * @return EntityInterface
      */
@@ -52,15 +78,14 @@ class BelongsTo extends AbstractRelation
     /**
      * Fetch data from foreign entity, when value of foreign key matches up with value of local key
      *
-     * @return EntityInterface|\Blast\Db\Entity\EntityInterface[]
+     * @return CollectionInterface
      */
     public function fetch()
     {
         $query = $this->getForeignEntity()->getMapper()->select();
         $result = $query->where(
-            $query->expr()->eq($this->getForeignKey(), $this->getEntity()->get($this->getLocalKey()))
-        )->setMaxResults(1)->execute(Query::RESULT_ENTITY);
-
+            $query->expr()->eq($this->getLocalKey(), $this->getForeignEntity()->get($this->getLocalKey()))
+        )->execute(Query::RESULT_COLLECTION);
         return $this->setResults($result)->getResults();
     }
